# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  dto.AllMyAttemptsResponse:
    properties:
      attempts:
        items:
          $ref: '#/definitions/dto.MyAttemptItem'
        type: array
      limit:
        type: integer
      page:
        type: integer
      statistics:
        $ref: '#/definitions/dto.AttemptsStatistics'
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.AnswerResultDetail:
    properties:
      correct_answer_id:
        type: string
      correct_answer_text:
        type: string
      is_correct:
        type: boolean
      point_earned:
        type: integer
      question_id:
        type: string
      question_number:
        type: integer
      question_point:
        type: integer
      question_title:
        type: string
      your_answer_id:
        type: string
      your_answer_text:
        type: string
    type: object
  dto.AttemptDetailResponse:
    properties:
      answers:
        items:
          $ref: '#/definitions/dto.AnswerResultDetail'
        type: array
      attempt_datetime:
        type: string
      attempt_id:
        type: string
      correct_answers:
        type: integer
      course_id:
        type: string
      course_name:
        type: string
      is_passed:
        type: boolean
      min_pass_score:
        type: integer
      quiz_id:
        type: string
      quiz_name:
        type: string
      result_summary:
        $ref: '#/definitions/dto.QuizResultSummary'
      score_achieved:
        type: integer
      total_questions:
        type: integer
    type: object
  dto.AttemptSummary:
    properties:
      attempt_datetime:
        type: string
      attempt_id:
        type: string
      attempt_number:
        type: integer
      correct_answers:
        type: integer
      is_passed:
        type: boolean
      score_achieved:
        type: integer
      total_questions:
        type: integer
    type: object
  dto.AttemptsStatistics:
    properties:
      average_score:
        type: number
      failed_attempts:
        type: integer
      passed_attempts:
        type: integer
      total_attempts:
        type: integer
      total_quizzes_attempted:
        type: integer
    type: object
...
  dto.CourseTypeInfo:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  dto.CourseTypeSummary:
    properties:
      id:
        type: string
      name:
        type: string
      quiz_count:
        type: integer
    type: object
...
  dto.MyAttemptItem:
    properties:
      attempt_datetime:
        type: string
      attempt_id:
        type: string
      correct_answers:
        type: integer
      course_id:
        type: string
      course_name:
        type: string
      is_passed:
        type: boolean
      min_pass_score:
        type: integer
      quiz_id:
        type: string
      quiz_name:
        type: string
      quiz_number:
        type: integer
      score_achieved:
        type: integer
      total_questions:
        type: integer
    type: object
...
  dto.MyQuizItem:
    properties:
      course_id:
        type: string
      course_name:
        type: string
      course_type:
        $ref: '#/definitions/dto.CourseTypeInfo'
      created_at:
        type: string
      id:
        type: string
      is_pass_required:
        type: boolean
      latest_attempt_date:
        type: string
      min_pass_score:
        type: integer
      my_attempts_count:
        type: integer
      my_best_score:
        type: integer
      name:
        type: string
      number:
        type: integer
      status:
        type: string
      thumbnail:
        type: string
      total_points:
        type: integer
      total_questions:
        type: integer
      updated_at:
        type: string
    type: object
  dto.MyQuizzesResponse:
    properties:
      completed_quizzes:
        type: integer
      course_types:
        items:
          $ref: '#/definitions/dto.CourseTypeSummary'
        type: array
      in_progress_quizzes:
        type: integer
      not_started_quizzes:
        type: integer
      quizzes:
        items:
          $ref: '#/definitions/dto.MyQuizItem'
        type: array
      total_quizzes:
        type: integer
    type: object
...
  dto.QuizAnswerForStudentResponse:
    properties:
      answer_text:
        type: string
      id:
        type: string
    type: object
  dto.QuizAttemptsHistoryResponse:
    properties:
      attempts:
        items:
          $ref: '#/definitions/dto.AttemptSummary'
        type: array
      best_score:
        type: integer
      is_completed:
        type: boolean
      latest_score:
        type: integer
      min_pass_score:
        type: integer
      quiz_id:
        type: string
      quiz_name:
        type: string
      total_attempts:
        type: integer
    type: object
  dto.QuizInCourseResponse:
    properties:
      course_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_completed:
        type: boolean
      is_pass_required:
        type: boolean
      min_pass_score:
        type: integer
      my_attempts_count:
        type: integer
      my_best_score:
        type: integer
      name:
        type: string
      number:
        type: integer
      thumbnail:
        type: string
      total_points:
        type: integer
      total_questions:
        type: integer
      updated_at:
        type: string
    type: object
  dto.QuizQuestionForStudentResponse:
    properties:
      answers:
        items:
          $ref: '#/definitions/dto.QuizAnswerForStudentResponse'
        type: array
      id:
        type: string
      number:
        type: integer
      point:
        type: integer
      question_title:
        type: string
    type: object
...
  dto.QuizResultSummary:
    properties:
      percentage:
        type: number
      points_earned:
        type: integer
      total_points:
        type: integer
    type: object
...
  dto.StartQuizResponse:
    properties:
      description:
        type: string
      is_pass_required:
        type: boolean
      min_pass_score:
        type: integer
      questions:
        items:
          $ref: '#/definitions/dto.QuizQuestionForStudentResponse'
        type: array
      quiz_id:
        type: string
      quiz_name:
        type: string
      total_points:
        type: integer
      total_questions:
        type: integer
    type: object
  dto.StudentAnswerInput:
    properties:
      answer_id:
        type: string
      question_id:
        type: string
    required:
    - answer_id
    - question_id
    type: object
...
  dto.SubmitQuizRequest:
    properties:
      answers:
        items:
          $ref: '#/definitions/dto.StudentAnswerInput'
        minItems: 1
        type: array
    required:
    - answers
    type: object
  dto.SubmitQuizResponse:
    properties:
      answers:
        items:
          $ref: '#/definitions/dto.AnswerResultDetail'
        type: array
      attempt_datetime:
        type: string
      attempt_id:
        type: string
      correct_answers:
        type: integer
      is_passed:
        type: boolean
      min_pass_score:
        type: integer
      quiz_id:
        type: string
      quiz_name:
        type: string
      result_summary:
        $ref: '#/definitions/dto.QuizResultSummary'
      score_achieved:
        type: integer
      total_questions:
        type: integer
    type: object
...
  utils.ErrorResponse:
    properties:
      errors: {}
      message:
        type: string
      status:
        example: false
        type: boolean
    type: object
  utils.StandardResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status:
        type: boolean
    type: object
host: 192.168.100.247:8080
info:
  contact: {}
  description: API documentation for Course Online
  title: Course Online API
  version: "1.0"
paths:
...
  /api/student/courses/{courseId}/quizzes:
    get:
      consumes:
      - application/json
      description: Retrieve all quizzes inside a specific course for the authenticated
        student.
      parameters:
      - description: Course ID (UUID format)
        format: uuid
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Quizzes retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.QuizInCourseResponse'
                  type: array
              type: object
        "400":
          description: Invalid course ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get quizzes for a course
      tags:
      - students' quiz answers
  /api/student/quizzes/{quizId}/attempts:
    get:
      consumes:
      - application/json
      description: Retrieve all attempts the student has made for a specific quiz.
      parameters:
      - description: Quiz ID (UUID format)
        format: uuid
        in: path
        name: quizId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Attempts retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.QuizAttemptsHistoryResponse'
              type: object
        "400":
          description: Invalid quiz ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Quiz not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get quiz attempt history
      tags:
      - students' quiz answers
  /api/student/quizzes/{quizId}/start:
    get:
      consumes:
      - application/json
      description: Initialize a quiz attempt and retrieve all questions with answer
        options.
      parameters:
      - description: Quiz ID (UUID format)
        format: uuid
        in: path
        name: quizId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Quiz started successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.StartQuizResponse'
              type: object
        "400":
          description: Invalid quiz ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Quiz not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Start a quiz
      tags:
      - students' quiz answers
  /api/student/quizzes/{quizId}/submit:
    post:
      consumes:
      - application/json
      description: Submit all answers to a quiz and receive scoring and result summary.
      parameters:
      - description: Quiz ID (UUID format)
        format: uuid
        in: path
        name: quizId
        required: true
        type: string
      - description: Submit quiz request payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubmitQuizRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Quiz submitted successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SubmitQuizResponse'
              type: object
        "400":
          description: Invalid request data or validation error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Quiz not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Submit quiz answers
      tags:
      - students' quiz answers
  /api/student/quizzes/attempts/{attemptId}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information of a specific quiz attempt, including
        answers and scoring.
      parameters:
      - description: Attempt ID (UUID format)
        format: uuid
        in: path
        name: attemptId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Attempt detail retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AttemptDetailResponse'
              type: object
        "400":
          description: Invalid attempt ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Attempt not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get quiz attempt detail
      tags:
      - students' quiz answers
  /api/student/quizzes/categories/{categoryId}/my-quizzes:
    get:
      consumes:
      - application/json
      description: Retrieve all quizzes from enrolled courses filtered by course type/category
        for the authenticated student.
      parameters:
      - description: Course Type/Category ID (UUID format)
        format: uuid
        in: path
        name: categoryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: My quizzes by category retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.MyQuizzesResponse'
              type: object
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get my quizzes filtered by course category
      tags:
      - students' quiz answers
  /api/student/quizzes/my-attempts:
    get:
      consumes:
      - application/json
      description: Retrieve all quiz attempts made by the current authenticated student
        with pagination.
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort order (asc|desc)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: My attempts retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AllMyAttemptsResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get all quiz attempts by the authenticated student
      tags:
      - students' quiz answers
  /api/student/quizzes/my-quizzess:
    get:
      consumes:
      - application/json
      description: Retrieve all quizzes from all enrolled courses for the authenticated
        student with progress tracking.
      produces:
      - application/json
      responses:
        "200":
          description: My quizzes retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.MyQuizzesResponse'
              type: object
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get all my quizzes
      tags:
      - students' quiz answers
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"